+++ test-composer-integration/portal-package/src/Portal.php
@@ -5,7 +5,14 @@ declare(strict_types=1);
 namespace HeptacomFixture\Portal\A;
 
 use Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalContract;
+use HeptacomFixture\Portal\A\Dto\NotEvenService;
 
 class Portal extends PortalContract
 {
+    public function getContainerExcludedClasses(): array
+    {
+        return [
+            NotEvenService::class,
+        ];
+    }
 }